Reverse the meaning of the flip-Y setting
If flip_Y is false, the Y axis of the world space is now in the "up"
direction. Therefore no flip -> the mathematical convention.

By default flip_y will be set to true as most of the test files have
the same orientation as the display, i.e. Y is in the "down" direction.