Set CMAKE_OSX_DEPLOYMENT_TARGET to 10.15
At the time of writing this commit the project is built for macOS
on the Github runners macos-latest, i.e. macos-14:

  https://github.com/actions/runner-images/blob/main/images/macos/macos-14-Readme.md

Those are M1 servers. The default XCODE version used is 15.4, which supports
the following range of deployment SDK: macOS 10.13-14.5. According to:

  https://developer.apple.com/support/xcode/

However since std::filesystem:path is only supported since macOS 10.15
we set our deployment version to 10.15.